+++ src/DataTableComponent.php
@@ -100,13 +100,6 @@ abstract class DataTableComponent extends Component
         throw new DataTableConfigurationException('You must either specify a model or implement the builder method.');
     }
 
-    /**
-     * The view to add any modals for the table, could also be used for any non-visible html
-     */
-    public function customView(): string
-    {
-        return 'livewire-tables::stubs.custom';
-    }
 
     public function render(): \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
     {