+++ CHANGELOG.md
@@ -9,7 +9,8 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 - Revert previous splitting of JS Files
 - Create additional Exception Classes (NoColumnsException, NoSearchableColumnsException, NoSortableColumnsException)
 - Fix footer rendering issue with extra td displayed depending on bulk action statuses
-- Move render data provision into WithData Trait
+- Create new WithCustomisations Trait
+- Move render data provision into Traits -> WithColumns, WithCustomisations, WithData
 
 ## [v3.1.3] - 2023-11-03
 - Add additional Lifecycle Hook by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1534